feat: implement afterSuccess function
if an afterSuccess function is defined on a CrudCollection, it will run
every time a request was successful. you can even return an observable
and the initial observable will wait for completion. this is handy for
creating script-ld tags on ssr for example, so the request waits until
this is done to resolve.